log_comm_impl
